// src/app/interfaces/instructor.ts
export interface Instructor {
    id_instructores: number;
    nombres: string;
    apellidos: string;
    dni: string;
    titulo: string;
    celular: string;
    foto_perfil: string;
    cv_corto: string;
    mp_public_key: string;
    mp_access_token: string;
    estado_eliminacion: boolean;
    id_usuarios: number;
    email: string;
}